ubuntu/upstart: run housekeeping script after 1 day
With systemd, we only set 'OnUnitActiveSec=1d' and 'OnBootSec' isn't set.
So housekeeping service will be run at the end of timer period.

In old Ubuntu we use upstart to run housekeeping script, this patch
changes the upstart to run housekeeping script after 1 day.

Just keep upstart to be consistent with systemd. In future, we may
change this according to feature requests.

Signed-off-by: Amos Kong <amos@scylladb.com>