redis: create redis keyspaces as configured
Currently we always try to create 16 keyspaces for Redis, the config
redis_default_database_count doesn't effect.

SELECT command is supported, it would switch between keyspaces. The limitation
is got from configuration. So if we config redis_default_database_count to be
larger than 16, SELECT allows to switch to unexisting keyspace, and it will
fail in read/write data.

Example:
| # scylla --developer-mode 1 --workdir ./ --enable-redis-protocol 1  --redis-default-database-count 20
| # redis-cli
| 127.0.0.1:6379> select 18
| OK
| 127.0.0.1:6379[18]> set a 1
| (error) ERR Can't find a column family STRINGs in keyspace REDIS_18
| 127.0.0.1:6379[18]> get a
| (error) ERR Can't find a column family STRINGs in keyspace REDIS_18

Signed-off-by: Amos Kong <amos@scylladb.com>